fix: resolve all build warnings for clean compilation
- Fix LNK4042: change draw_game_overlay.h from ClCompile to ClInclude in vcxproj
- Fix C4244: add static_cast<int>() for int64_t narrowing in postProcess, trt_detector, nvinf
- Fix C4566: remove unicode narrow no-break space from virtual_camera.cpp
- Fix C4005: add #ifndef guard for KEY_EXECUTE macro in HidTable.h
- Fix C4244: use proper WideCharToMultiByte for wchar_t to string conversion in dml_detector

+++ detector/postProcess.cpp
@@ -150,7 +150,7 @@ std::vector<Detection> postProcessYolo11(
         return detections;
     }
     
-    cv::Mat det_output(rows, cols, CV_32F, (void*)output);
+    cv::Mat det_output(static_cast<int>(rows), static_cast<int>(cols), CV_32F, (void*)output);
 
     const float img_scale = trt_detector.img_scale;
     
@@ -246,7 +246,7 @@ std::vector<Detection> postProcessYolo11DML(
     int64_t rows = shape[0];
     int64_t cols = shape[1];
 
-    cv::Mat det_output(rows, cols, CV_32F, (void*)output);
+    cv::Mat det_output(static_cast<int>(rows), static_cast<int>(cols), CV_32F, (void*)output);
     for (int i = 0; i < cols; ++i) {
         cv::Mat classes_scores = det_output.col(i).rowRange(4, 4 + numClasses);
         cv::Point class_id_point;
